Modified code to match changes in the Score class
Changed line 52 from HashMap<String, Integer> to HashMap<String, Double>

+++ src/GameViewer.java
@@ -49,7 +49,7 @@ public interface GameViewer {
     /**
      * @return the finalScore
      */
-    HashMap<String, Integer> getFinalScore();
+    HashMap<String, Double> getFinalScore();
 
     /**
      * @return the gui
